Added swin earth padding
